Bug 1950375 - Remove exclusion for Windows when handling symlinks while dealing with PreprocessedFile. r=firefox-build-system-reviewers,sergesanspaille
We used to not use symbolink links on Windows, but that changed a while
back. There was a leftover exclusion causing problems when replacing a
symbolic link with a preprocessed file.

Differential Revision: https://phabricator.services.mozilla.com/D239860

UltraBlame original commit: 1a93acc599c43c59c6056e8945623f2b612a4b4d

+++ python/mozbuild/mozpack/files.py
@@ -551,10 +551,8 @@ class PreprocessedFile(BaseFile):
         
         
         
-        
-        if hasattr(os, "symlink") and platform.system() != "Windows":
-            if os.path.islink(dest.path):
-                os.remove(dest.path)
+        if hasattr(os, "symlink") and os.path.islink(dest.path):
+            os.remove(dest.path)
 
         pp_deps = set(self.extra_depends)
